lib/metric-config-parser/pyproject.toml
[build-system]
requires = ["pytest-runner", "setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mozilla-metric-config-parser"
version = "2025.9.2"
authors = [{ name = "Mozilla Corporation", email = "fx-data-dev@mozilla.org" }]
description = "Parses metric configuration files."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "attrs",
  "cattrs",
  "Click",
  "GitPython",
  "jinja2",
  "mozilla-nimbus-schemas>=2024.9.3",
  "pytz",
  "requests",
  "toml",
]

[project.scripts]
metric-config-parser = "metric_config_parser.cli:cli"

[project.urls]
Repository = "https://github.com/mozilla/metric-hub/blob/main/lib/metric-config-parser"
Documentation = "https://github.com/mozilla/metric-hub/blob/main/lib/metric-config-parser"

[project.optional-dependencies]
testing = [
  "coverage",
  "jsonschema",
  "mypy",
  "pytest",
  "pytest-cov",
  "pytest-ruff",
  "ruff",
  "types-futures",
  "types-protobuf",
  "types-pytz",
  "types-PyYAML",
  "types-requests",
  "types-setuptools",
  "types-six",
  "types-toml",
]

[tool.setuptools]
packages = [
  "metric_config_parser",
  "metric_config_parser.tests",
  "metric_config_parser.tests.integration",
]
[tool]
[tool.setuptools.package-data]
metric_config_parser = ["templates/*"]
"metric_config_parser.tests" = ["data/*"]

[tool.ruff]
line-length = 100
exclude = [".git", "__pycache__"]
target-version = "py310"

[tool.ruff.lint]
ignore = ["E741", "RUF005", "RUF012", "SIM105"]
select = [
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "E",   # pycodestyle
  "F",   # Pyflakes
  "I",   # isort
  "PT",  # flake8-pytest-style
  "Q",   # flake8-quotes
  "RUF", # Ruff-specific rules
  "SIM", # flake8-simplify
  "TC",  # flake8-type-checking
  "TID", # flake8-tidy-imports
  "UP",  # pyupgrade
  "W",   # pycodestyle
]
